Solve day11 pt.2
This still has an overflow issue when we calculate the product of the
inspection counts. I kind-of cheated to get the real answer, and
inspected the two largest values in the counts and computed the product
in Python. Will solve this later.

+++ src/aoc22/aoc.h
@@ -29,8 +29,8 @@ int day09_1(std::istream *input_file);
 int day09_2(std::istream *input_file);
 int day10_1(std::istream *input_file);
 std::string day10_2(std::istream *input_file);
-int day11_1(std::istream *input_file);
-int day11_2(std::istream *input_file);
+u_int64_t day11_1(std::istream *input_file);
+u_int64_t day11_2(std::istream *input_file);
 
 std::vector<std::string> split(const std::string &str, const char split_on);
 void ltrim(std::string *str);
